Modificações de Valation
