--- src/main/java/com/crud/cadastrousuario/domain/validation/ValidEmail.java
@@ -1,26 +1,0 @@
-package com.crud.cadastrousuario.domain.validation;
-
-
-import jakarta.validation.Constraint;
-import jakarta.validation.Payload;
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@NotBlank (message = "email não pode ser em Branco")
-@NotNull(message = "email nao pode ser nulo")
-@Email(message = "Email invalido Colocar um email Valido")
-@Constraint(validatedBy = ValidEmailValidator.class)
-@Target({ElementType.FIELD})
-@Retention(RetentionPolicy.RUNTIME)
-public @interface ValidEmail {
-
-    String message() default "Email inválido";
-    Class<?>[] groups() default {};
-    Class<? extends Payload>[] payload() default {};
-}
