+++ src/main/java/com/crud/cadastrousuario/exemplos/validation/ValidEmail.java
@@ -1,28 +1,26 @@
-package com.crud.cadastrousuario.domain.validation;
+package com.crud.cadastrousuario.exemplos.validation;
+
 
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
+import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
 
 import java.lang.annotation.ElementType;
-import java.lang.annotation.Documented;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-//@NotNull(message = "Nome não pode ser nulo")
-//@NotBlank(message = "Nome não pode ser vazio")
-//@Size(min = 3, max = 50, message = "O número de caracteres deve estar entre 3 e 50 caracteres")
-@Constraint(validatedBy = ValidNameValidator.class)
+@NotBlank (message = "email não pode ser em Branco")
+@NotNull(message = "email nao pode ser nulo")
+@Email(message = "Email invalido Colocar um email Valido")
+@Constraint(validatedBy = ValidEmailValidator.class)
 @Target({ElementType.FIELD})
 @Retention(RetentionPolicy.RUNTIME)
-public @interface ValidName {
+public @interface ValidEmail {
 
-    String message() default "Nome inválido";
+    String message() default "Email inválido";
     Class<?>[] groups() default {};
     Class<? extends Payload>[] payload() default {};
-
 }
-
