+++ src/main/java/com/crud/cadastrousuario/exemplos/validation/ValidNameValidator.java
@@ -1,11 +1,13 @@
-package com.crud.cadastrousuario.domain.validation;
+package com.crud.cadastrousuario.exemplos.validation;
+
 
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 
-public class ValidEmailValidator implements ConstraintValidator<ValidEmail, String> {
-    @Override
-    public void initialize(ValidEmail constraintAnnotation) {
+public class ValidNameValidator implements ConstraintValidator<ValidName, String> {
+
+    public void initialize(ValidName constraintAnnotation) {
+
         ConstraintValidator.super.initialize(constraintAnnotation);
     }
 
@@ -14,17 +16,17 @@ public class ValidEmailValidator implements ConstraintValidator<ValidEmail, Stri
         context.disableDefaultConstraintViolation();
 
         if (value == null) {
-            context.buildConstraintViolationWithTemplate("Email não pode ser nulo")
+            context.buildConstraintViolationWithTemplate("Nome não pode ser nulo")
                     .addConstraintViolation();
             return false;
         }
         if (value.trim().isEmpty()) {
-            context.buildConstraintViolationWithTemplate("Email não pode ser em branco")
+            context.buildConstraintViolationWithTemplate("Nome não pode ser vazio")
                     .addConstraintViolation();
             return false;
         }
-        if (!value.contains("@") || !value.contains(".")) {
-            context.buildConstraintViolationWithTemplate("Email inválido. Colocar um email válido")
+        if (value.trim().length() < 3 || value.trim().length() > 50) {
+            context.buildConstraintViolationWithTemplate("O número de caracteres deve estar entre 3 e 50 caracteres")
                     .addConstraintViolation();
             return false;
         }
